Add initial henshin parse transform
This parse transform checks whether any module in a remote function call
is an henshin module and apply any transformation rule found to the
given AST.

A warning is issued if a module cannot be find at compile-time.

Care is taken to handle errors gracefully.

When a rule is applied, its result is transformed so that any introduced
rule call is also transformed.